% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chebpol-deprecated.R
\name{chebpol-deprecated}
\alias{chebpol-deprecated}
\title{Deprecated functions}
\description{
As of version 1.7, the functions for creating various interpolants
have been deprecated. These are \code{chebappx, chebappxf, chebappxg,
chebappxgf, fhappx, mlappx, slappx, ucappx, ucappxf, polyh}. The wrapper \code{ipol}
should be used instead. The wrapper gives a uniform interface.
}
