% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{havefftw}
\alias{havefftw}
\title{Check whether chebpol uses FFTW}
\usage{
havefftw()
}
\value{
Returns TRUE if chebpol uses FFTW. Otherwise FALSE.
}
\description{
It is possible to compile chebpol without FFTW.  If this is done, it will
not be feasible with high-degree Chebyshev polynomials.  I.e. a 100x100x100
approximation will be possible, but not a one-dimensional 1000000.  This
function checks whether chebpol uses FFTW.
}
