% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexlinear.R
\name{slappx}
\alias{slappx}
\title{Simplex Linear approximation on scattered data}
\usage{
slappx(val, knots, ...)
}
\arguments{
\item{val}{Array or function. Function values in the knots, or the function
itself.}

\item{knots}{matrix. Each column is a point in M-dimensional space.}

\item{...}{Further arguments to the function, if \code{is.function(val)}.}
}
\value{
A \code{function(x)} interpolating the values.
}
\description{
A call \code{fun <- slappx(val, knots)} creates a piecewise multilinear interpolation
on the Delaunay triangulation of the \code{knots}.
}
\note{
The simplex linear interpolation uses package \pkg{geometry} heavily, it is not 
optimized for dimensions larger than 2, so it is quite slow, in particular evaluation of the
interpolant. It does not yield values for points outside the convex hull of the knots.
}
\examples{
knots <- matrix(runif(3*100), 3)
f <- function(x) exp(-sum(x^2))
g <- slappx(f, knots)
a <- matrix(runif(3*5), 3)
rbind(true=apply(a,2,f), sl=g(a))

}
